Adding .cbl, .bat, and .gitignore files.
This version is written for GNUCOBOL 2.0. The performance is quite poor
in Windows. Possibly that can be tweaked, but most likely I'll commit
ACUCOBOL versions from here on out. GNUCOBOL has the X"F4" and X"F5"
library routines, so it has no need of procedures to do the same.